refactor: decouple panels from axes 02 - shuffled around other files
